+++ crates/gen/src/types/mod.rs
@@ -1,5 +1,7 @@
 pub use super::*;
 
+mod bool32;
+mod bstr;
 mod callback;
 mod class;
 mod com_interface;
@@ -8,8 +10,18 @@ mod delegate;
 mod r#enum;
 mod function;
 mod interface;
+mod matrix3x2;
+mod matrix4x4;
+mod pstr;
+mod pwstr;
 mod r#struct;
+mod timespan;
+mod vector2;
+mod vector3;
+mod vector4;
 
+pub use bool32::*;
+pub use bstr::*;
 pub use callback::*;
 pub use class::*;
 pub use com_interface::*;
@@ -17,5 +29,13 @@ pub use constant::*;
 pub use delegate::*;
 pub use function::*;
 pub use interface::*;
+pub use matrix3x2::*;
+pub use matrix4x4::*;
+pub use pstr::*;
+pub use pwstr::*;
 pub use r#enum::*;
 pub use r#struct::*;
+pub use timespan::*;
+pub use vector2::*;
+pub use vector3::*;
+pub use vector4::*;
